==> analyses/01-data-wrangling.Rmd <==
---
title: "Preparing the data and tree for analysis"
author: "Natalie Cooper"
output: html_document
---

## Preparation

```{r, message = FALSE}
# Load libraries
library(tidyverse)
library(here)
library(ape)
library(phytools)
```

Read in the data. This is the data from Marcello
```{r, message = FALSE}
ds <- read_csv(here("raw-data/CHROMO_STATES_2020-10-16.csv"))
```

Take a look at the dataset
```{r}
#glimpse(ds)
head(ds)
```

Add the haploid number of chromosomes which is needed for some later analyses
```{r}
ds <-
  ds %>%
  mutate(haploidn = chromon/2)
```

Fix up species names in our data to match them to the ecological data. Note that this will give a warning saying that some of the species names have more than two words in them, this is fine as we don't need the extra bits.
```{r, message = FALSE}
ds <- ds %>%
  # remove outgroup
  # filter(species != "leiolepis_belliana") %>%
  # capitalise genus column
  mutate(genus = capitalize(genus)) %>%
  # split the species column into Genus and Species
  separate(species, into = c("Genus", "Species"), sep = "_", remove = FALSE) %>%
  # capitalise Genus column
  mutate(Genus = capitalize(Genus)) %>%
  # stick Genus and Species together as Binomial
  unite("Binomial", Genus, Species, sep = " ") %>%
  # Create a new column for the things MM uses as OTUs
  mutate(Binomial_Marcello = Binomial) %>%
  # Correct the taxonomy so it matches Uetz
  mutate(Binomial = case_when(Binomial == "Brookesia thieli1" ~ "Brookesia thieli",
                              Binomial == "Brookesia thieli2" ~ "Brookesia thieli",
                              Binomial == "Brookesia ramantsoai" ~ "Brookesia ramanantsoai",
                              Binomial == "Brookesia tubercolata" ~ "Brookesia tuberculata",
                              Binomial == "Furcifer verrucosus1" ~ "Furcifer verrucosus",
                              Binomial == "Furcifer verrucosus2" ~ "Furcifer verrucosus",
                              Binomial == "Furcifer belalandensis" ~ "Furcifer belalandaensis",
                              Binomial == "Furcifer wilsii" ~ "Furcifer willsii",
                              Binomial == "Calumma malthae" ~ "Calumma malthe",
                              Binomial == "Calumma brevicorne1" ~ "Calumma brevicorne",
                              Binomial == "Calumma brevicorne2" ~ "Calumma brevicorne",
                              Binomial == "Calumma gastrotaenia2" ~ "Calumma gastrotaenia",
                              Binomial == "Calumma emelinae" ~ "Calumma emelinae",
                              Binomial == "Calumma fallax1" ~ "Calumma fallax",
                              Binomial == "Calumma fallax2" ~ "Calumma fallax",
                              Binomial == "Calumma gallus1" ~ "Calumma gallus",
                              Binomial == "Calumma gallus2" ~ "Calumma gallus",
                              Binomial == "Calumma gallus3" ~ "Calumma gallus",
                              Binomial == "Calumma boettegeri" ~ "Calumma boettgeri",
                              Binomial == "Bradypodion pumilium" ~ "Bradypodion pumilum",
                              Binomial == "Tioceros hoehnelii" ~ "Trioceros hoehnelii",
                              Binomial == "Trioceros jacksoni" ~ "Trioceros jacksonii",
                              Binomial == "Brookesia superciliaris2" ~ "Brookesia superciliaris",
                              TRUE ~ as.character(Binomial)))

```

Add the ecological data from Meiri 2018 and match up to the chromosome data. This prints out the species there are no matching data for in Meiri 2018 (should be 7 species). These species remain in the dataset, but don't have matching ecological data.
```{r, message = FALSE}
# Read in the Meiri data
eco <- read_csv(here("raw-data/Meiri2018_AppendixS1.csv"))
#head(eco)

# Extract only chameleons and only relevant columns
eco <- eco %>%
  filter(Family == "Chamaeleonidae") %>%
  select(Binomial, Realm = `main biogeographic Realm`,
         Latitude = `Latitude centroid (from Roll et al. 2017)`, 
         Longitude = `Longitude centroid (from Roll et al. 2017)`, 
         maxSVL = `maximum SVL`,     
         Substrate = substrate, 
         Repromode = `reproductive mode`, 
         minClutchSize = `smallest clutch`, 
         maxClutchSize =  `largest clutch`, 
         minBreedingAge = `youngest age at first breeding (months)`, 
         maxBreedingAge =  `oldest age at first breeding (months)`)

# What species are missing?
setdiff(ds$Binomial, eco$Binomial)

# Merge
ds <- left_join(ds, eco)

# Rename sample to tips
mydata <- rename(ds, tips = sample)
```

Read in the tree, then look at the tree structure and check a few features
```{r}
tree <- read.tree(here("data/Bayesian-tree.tre"))
```

```{r}
# Look at tree structure
str(tree)

# Check tree is binary
is.binary(tree) # TRUE
# Check tree is rooted
is.rooted(tree) # TRUE

# Check tree is ultrametric
is.ultrametric(tree) # TRUE
```

Create a species level tree. 
Five species have OTUs with different numbers of chromosomes, so these should be retained. Otherwise collapse OTUs to species at tips
```{r}
mydata2 <- mydata %>%
          mutate(Binomial = case_when(species == "calumma_gallus1" ~ "Calumma gallus1",
                                      species == "calumma_gallus2" ~ "Calumma gallus2",
                                      species == "calumma_gallus3" ~ "Calumma gallus3",
                                      species == "calumma_fallax1" ~ "Calumma fallax1",
                                      species == "calumma_fallax2" ~ "Calumma fallax2",
                                      species == "furcifer_verrucosus1" ~ "Furcifer verrucosus1",
                                      species == "furcifer_verrucosus2" ~ "Furcifer verrucosus2",
                                      species == "calumma_brevicorne1" ~ "Calumma brevicorne1",
                                      species == "calumma_brevicorne2" ~ "Calumma brevicorne2",
                                      species == "calumma_parsonii" ~ "Calumma parsonii1",
                                      species == "calumma_parsonii_cristifer" ~ "Calumma parsonii2",
                                      TRUE ~ as.character(Binomial)))

# Select the first row for each species
mydata2 <- 
  mydata2 %>%
  group_by(Binomial) %>%
  filter(row_number()==1)

mytree2 <- drop.tip(tree, setdiff(tree$tip.label, mydata2$tips))
```

Finally write these cleaned data out for use in later analyses.

Should be 93 taxa in the "species" analyses
```{r}
write_csv(mydata, file = here("data/chromosome-data.csv"))
#Species level
write_csv(mydata2, file = here("data/chromosome-data-species.csv"))
write.tree(mytree2, file = here("data/Bayesian-tree-species.tre"))
```
