==> analyses/03B-visualising-data-on-phylogenies_OTUs.Rmd <==
---
title: "Visualising the data on the phylogeny"
author: "Natalie Cooper"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message = FALSE}
# Load libraries
library(tidyverse)
library(here)
library(phytools)
library(ggtree)
library(ggnewscale)
library(patchwork)
```

Read in the data and the tree
```{r, message = FALSE, echo = FALSE}
mydata <- read_csv(here("data/chromosome-data.csv"))
mytree <- read.tree(here("data/Bayesian-tree.tre"))

# look at the dataset and tree
#str(mytree)
#head(mydata)
```

Extract some subsets of the data for the plotting
```{r}
# chromon
chromon <-
  mydata %>%
  select(haploidn) %>%
  as.data.frame()
# Add rownames as tip names  
rownames(chromon) <- mydata$tips

# armn
arms <-
  mydata %>%
  select(armn) %>%
  as.data.frame()
# Add rownames as tip names  
rownames(arms) <- mydata$tips

# macro pairs
macropairs <-
  mydata %>%
  select(macrop) %>%
  as.data.frame()
# Add rownames as tip names  
rownames(macropairs) <- mydata$tips

# micro pairs
micropairs <-
  mydata %>%
  select(microp) %>%
  as.data.frame()
# Add rownames as tip names  
rownames(micropairs) <- mydata$tips
```

Set up colour palettes

```{r}
mycolours <- c("#000000","#208EA3", "#FD817D", "#4178BC", "#EA4E9D",
                "#EECC16", "#E8384F", "#E37CFF", "#37A862","#A4C61A","#7A71F6")
```

## Visualising data on the tree

Make trees using ggtree. Note that this code will throw lots of warnings about scales which can be ignored
```{r}
# Make the circular tree base
circ <- ggtree(mytree, layout = "circular")

# Add genera
genus_tree <-
  circ +
  geom_cladelabel(node = MRCA(circ, mydata$tips[mydata$genus == "Bradypodion"]), 
                  label = expression(italic("Bradypodion")), 
                align =FALSE, offset = 1, fontsize = 4, offset.text = 60) +
  geom_cladelabel(node = MRCA(circ, mydata$tips[mydata$genus == "Brookesia"]),
                  label = expression(italic("Brookesia")), 
                align = TRUE, offset = 1, fontsize = 4, offset.text = 10) +
  geom_cladelabel(node = MRCA(circ, mydata$tips[mydata$genus == "Calumma"]),
                  label = expression(italic("Calumma")), 
                align = TRUE, offset = 1, fontsize = 4, offset.text = 10) +
  geom_cladelabel(node = MRCA(circ, mydata$tips[mydata$genus == "Chamaeleo"]),
                  label = expression(italic("Chamaeleo")), 
                align = TRUE, offset = 1, fontsize = 4, offset.text = 40) +
  geom_cladelabel(node = MRCA(circ, mydata$tips[mydata$genus == "Furcifer"]), 
                  label = expression(italic("Furcifer")), 
                align = TRUE, offset = 1, fontsize = 4, offset.text = 55) +
  geom_cladelabel(node = MRCA(circ, mydata$tips[mydata$genus == "Kinyongia"]),
                  label = expression(italic("Kinyongia")), 
                align = TRUE,offset = 1, fontsize = 4, offset.text = 25) +
  geom_cladelabel(node = MRCA(circ, mydata$tips[mydata$genus == "Palleon"]), 
                  label = expression(italic("Palleon")), 
                align = TRUE,offset = 1, fontsize = 4, offset.text = 10) +
  geom_cladelabel(node = MRCA(circ, mydata$tips[mydata$genus == "Rhampholeon"]),
                  label = expression(italic("Rhampholeon")), 
                align = TRUE,offset = 1, fontsize = 4, offset.text = 10) +
  geom_cladelabel(node = MRCA(circ, mydata$tips[mydata$genus == "Trioceros"]),
                  label = expression(italic("Trioceros")), 
                align = TRUE, offset = 1, fontsize = 4, offset.text = 45)

# Save tree
# ggsave(genus_tree, filename = here("outputs/OTUs/trees-genera-bw.png"), height = 5)
```

```{r}
grp2 <- list(Bradypodion = c(mydata$tips[mydata$genus == "Bradypodion"]), 
            Brookesia = c(mydata$tips[mydata$genus == "Brookesia"]), 
            Calumma = c(mydata$tips[mydata$genus == "Calumma"]), 
            Chamaeleo = c(mydata$tips[mydata$genus == "Chamaeleo"]), 
            Furcifer = c(mydata$tips[mydata$genus == "Furcifer"]), 
            Kinyongia = c(mydata$tips[mydata$genus == "Kinyongia"]), 
            Palleon = c(mydata$tips[mydata$genus == "Palleon"]), 
            Rhampholeon = c(mydata$tips[mydata$genus == "Rhampholeon"]),
            Rieppeleon = c(mydata$tips[mydata$genus == "Rieppeleon"]),
            Trioceros = c(mydata$tips[mydata$genus == "Trioceros"])) 

groupOTU(genus_tree, grp2, 'Genus') + aes(color = Genus) +
  scale_colour_manual(values = mycolours) +
  theme(legend.position="none")

# Save tree
# ggsave(filename = here("outputs/OTUs/trees-genera-colour.png"), height = 5)
```

```{r}
# Without labels but with a legend instead
groupOTU(circ, grp2, 'Genus') + aes(color = Genus) +
  theme(legend.position="right") +
  scale_colour_manual(values = mycolours, name = "", 
                      breaks = sort(unique(mydata$genus)),
                      labels = c(expression(italic("Bradypodion")),
                                 expression(italic("Brookesia")),
                                 expression(italic("Calumma")),
                                 expression(italic("Chamaeleo")),
                                 expression(italic("Furcifer")),
                                 expression(italic("Kinyongia")),
                                 expression(italic("Palleon")),
                                 expression(italic("Rhampholeon")),
                                 expression(italic("Rieppeleon")),
                                 expression(italic("Trioceros")))) +
  theme_void(base_size = 14) +
  theme(legend.text.align = 0)

# Save tree
# ggsave(filename = here("outputs/OTUs/trees-genera-colour-legend.png"), height = 5)
```

Add the chromosome data...
```{r}
# Plot trees
p1 <- 
  gheatmap(circ, chromon, offset =.8, width =.2, colnames = FALSE) +
  scale_fill_viridis_c(na.value = "grey90", option = "plasma", name = "") +
  ggtitle("chromosomes (n)") +
  theme(plot.title = element_text(hjust = 0.5))
# Look at this
#p1

p2 <- 
  gheatmap(circ, arms, offset = .8, width =.2, colnames = FALSE) +
  scale_fill_viridis_c(na.value = "grey90", option = "magma", name = "") +
  ggtitle("arm number") +
  theme(plot.title = element_text(hjust = 0.5))
# Look at it
#p2

p3 <- 
  gheatmap(circ, macropairs, offset = .8, width =.2, colnames = FALSE) +
  scale_fill_viridis_c(na.value = "grey90", name = "") +
  ggtitle("macrochromosome pairs") +
  theme(plot.title = element_text(hjust = 0.5))
# Look at it
#p3

p4 <- 
  gheatmap(circ, micropairs, offset = .8, width =.2, colnames = FALSE) +
  scale_fill_viridis_c(na.value = "grey90", name = "") +
  ggtitle("microchromosome pairs") +
  theme(plot.title = element_text(hjust = 0.5))
# Look at it
#p4

(p1 + p2) / (p3 + p4)

# Save the plot
#ggsave(filename = here("outputs/OTUs/trees-chromosome-properties.png"))

```
