==> analyses/06C-figures-simulation-models_main.Rmd <==
---
title: "Simulations of chromosome numbers figures"
author: "Natalie Cooper"
output: html_document
---

## Looking at phylogenetic patterns using simulations

```{r, message = FALSE}
# Load libraries
library(tidyverse)
library(here)
library(phytools)
source(here("chromEvol/ChromEvol.R")) #install this from https://www.en.sysbot.bio.lmu.de/people/employees/cusimano/use_r/
library(patchwork)
```

Add the observed data to the simulations data, then work out the differences between observed numbers and simulation numbers of chromosomes.
```{r}
# Read in simulation output
sim_output <- read_csv(here("data/chromevol-simulations-outputs-best.csv"))

# Read in observed data
mydata <- read_csv(here("data/chromosome-data-species.csv"))

# Remove Rieppeleon and outgroup
mydata <- 
  mydata %>% 
  filter(genus != "Rieppeleon" & genus != "Leiolepis")

# Rename the haploidn column
# Select only tip names, genus and observed haploidn
mydata <-
  mydata %>%
  rename(obs_haploidn = haploidn) %>%
  dplyr::select(tips, obs_haploidn, genus)

# Join data by tip names
sim_output2 <- 
  full_join(sim_output, mydata, by = "tips")

# Calculate differences
results <- 
  sim_output2 %>%
  mutate(haploidn = as.numeric(haploidn)) %>%
  mutate(diff = obs_haploidn - haploidn)
```

Read in null models. Add the data so the differences between observed and simulated chromosome numbers can be calculated.
```{r}
# Read in null simulations
sims_output_null <- read_csv(here("data/null-simulations-outputs.csv"))

# Join data by tip names
sim_output_null2 <- 
  full_join(sims_output_null, mydata, by = "tips")

# Calculate differences
results_null <- 
  sim_output_null2 %>%
  mutate(diff = obs_haploidn - sim_haploidn)
```

## Plot the observed versus the simulations for both sets
```{r}
ggplot(results, aes(x = diff)) +
  geom_density(colour = "cornflowerblue") +
  geom_density(data = results_null, colour = "grey") +
  geom_vline(xintercept = 0, linetype = "dotted") +
  facet_wrap(~genus) +
  theme_bw(base_size = 14) +
  theme(strip.background = element_rect(fill = "white"), 
        strip.text = element_text(face = "italic"),
        axis.text.x = element_text(angle = 45, hjust = 1)) +
  xlab(expression(n[observed] - n[simulated]))

#ggsave(filename = here("outputs/all-simulations-differences-best.png"))  

```
Plot number of chromosomes predicted using chromosome evolution models
```{r}
ggplot(results_null, aes(x = sim_haploidn)) +
  # Add null as this is the true distribution
  geom_histogram(binwidth = 1, alpha = 0.5) +
  # Add simulations in blue
  geom_histogram(data = results, aes(x = haploidn), 
                 fill = "cornflowerblue", binwidth = 1, alpha = 0.5) +
  theme_bw(base_size = 14) +
  xlab("chromosomes (n)") +
  coord_cartesian(xlim = c(5, 25), ylim = c(0,40000))

#ggsave(filename = here("outputs/chromevol-simulations-numbers-best.png"), height = 5)  
```
